fix(admin/reports): move client code to ReportsClient and force-dynamic to avoid prerender/env issues
